Call an Rcpp function within a doParallel call

Constructs an example showing how to use foreach, iterators, and doParallel to perform a parallel computation with a C++ function written using Rcpp.

Usage

mean_parallel_compute(
  n,
  mean = 0,
  sd = 1,
  n_sim = 1000,
  n_cores = parallel::detectCores()
)

Arguments

n

Number of Observations

mean

Center of Normal Distribution

sd

Standard Deviation of Normal Distribution

n_sim

Number of Simulations to Run

n_cores

Number of CPU cores to use in parallelization task.

Value

A vector of length n_sim containing the mean for each distribution.

Details

The mean_parallel_compute() function performs a bootstrap computation in parallel of a mean value from the normal distribution.

Examples

# Compute the mean on 1000 observations with 50 replications across
# 2 CPUs.
mean_parallel_compute(1000, n_sim = 50, n_cores = 2)
#>                    [,1]
#> result.1  -0.0382227351
#> result.2   0.0175640263
#> result.3  -0.0031693617
#> result.4  -0.0294290384
#> result.5  -0.0475833715
#> result.6  -0.0017675815
#> result.7  -0.0013198371
#> result.8   0.0018643164
#> result.9   0.0099135244
#> result.10  0.0069427795
#> result.11 -0.0198964325
#> result.12  0.0169600905
#> result.13  0.0276053127
#> result.14  0.0352570315
#> result.15 -0.0062964592
#> result.16  0.0265083221
#> result.17  0.0007933019
#> result.18  0.0167588181
#> result.19 -0.0106756039
#> result.20 -0.0837588237
#> result.21  0.0396804132
#> result.22  0.0587824692
#> result.23  0.0354617554
#> result.24 -0.0568306561
#> result.25 -0.0067532291
#> result.26  0.0045832939
#> result.27 -0.0249538461
#> result.28 -0.0235042564
#> result.29 -0.0564515114
#> result.30  0.0452349411
#> result.31 -0.0015438869
#> result.32  0.0626669064
#> result.33 -0.0046249876
#> result.34  0.0116040449
#> result.35  0.0356202926
#> result.36 -0.0212270698
#> result.37 -0.0089656852
#> result.38  0.0223898616
#> result.39  0.0041604304
#> result.40 -0.0087430856
#> result.41  0.0234387442
#> result.42 -0.0228969288
#> result.43  0.0464021609
#> result.44  0.0126897430
#> result.45 -0.0177316694
#> result.46 -0.0191212452
#> result.47 -0.0132761246
#> result.48 -0.0220821433
#> result.49 -0.0085231496
#> result.50 -0.0245882887